requirements_personal.txt
Django==2.1.1
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
Pillow==6.1.0
pytz==2019.1
